+++ src/param.h
@@ -68,6 +68,7 @@ class Param {
     directly_called_ = other.directly_called_;
     help_ = other.help_;
     version_ = other.version_;
+    print_model_ = other.print_model_;
     std::swap(argv_vec_, other.argv_vec_);
   }
 
@@ -81,6 +82,7 @@ class Param {
     directly_called_ = other.directly_called_;
     help_ = other.help_;
     version_ = other.version_;
+    print_model_ = other.print_model_;
     std::swap(argv_vec_, other.argv_vec_);
     return *this;
   }
@@ -91,6 +93,7 @@ class Param {
     this->set_help(false);
     this->set_version(false);
     this->set_precision(6);
+    this->set_print_model(false);
     argc_i = 0;
   }
 
@@ -99,13 +102,16 @@ class Param {
   bool version() const { return version_; }
   bool read_init_genealogy() const { return this->read_init_genealogy_; }
   size_t random_seed() const { return random_seed_; }
+  size_t precision() const { return precision_; }
+  bool seed_is_set() const { return this->seed_set_; }
+  bool print_model() const { return this->print_model_; }
+
+  void set_precision ( const size_t p ) { this->precision_ = p; }
   void set_random_seed(const size_t seed) { 
     this->random_seed_ = seed;
     this->seed_set_ = true; 
   }
-  size_t precision() const { return precision_; }
-  void set_precision ( const size_t p ) { this->precision_ = p; }
-  bool seed_is_set() const { return this->seed_set_; }
+  void set_print_model(const bool print_model) { print_model_ = print_model; }
 
   // Other methods
   void printHelp(std::ostream& stream);
@@ -156,6 +162,7 @@ class Param {
   bool help_;
   bool version_;
   bool read_init_genealogy_;
+  bool print_model_;
   std::vector<char*> argv_vec_;
 };
 #endif