+++ NEWS.md
@@ -1,14 +1,23 @@
 scrm Version History
 ========================
 
+scrm 1.5.0
+------------------------
+Released: Not yet
+
++ New feature: Added flag "--print-model" which prints a textual representation
+  of the demographic model for verification and debugging. 
+
+
+
 scrm 1.4.1
 ------------------------
 Released: 2015-04-04
 
 + Bug fix: Wrong population size where calculated when migration rates
   changes or population splits and merges occurred in an growth period (#56)
-+ Added an error message when a population size is set to 0 (#52)
 + Bug fix: When -es was used at time 0, _scrm_ ran into an endless loop (#53)
++ Improvement: Added an error message when a population size is set to 0 (#52)